Workaround for the too many token problem
There are too many `token` keywords in the parser. The latest offender
is `identl`.

It makes things like that parse :

    $a :: b
    :: c

As

    $a::b::c

I just added a "try" in the `qualif` function to at least being able to
backtrack from such catastrophic groupings. A rewrite of the relevant
part would be required, but would require to be *very* careful.