Replace Data.FileCache with a plain cache with no watch
Watching for file changes it useless in the current client only setting.

If a server settings is set up, it would be nice to change the filecache package
to work with OS X. That would anyway require changing the APIs because the
underline fsnotify package is using a 'withXXX' pattern.